requirements_client.txt
python==3.9.1
aniso8601==8.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorzero==1.1
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SocketIO==4.2.1
gpiozero==1.5.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pigpio==1.44
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.16.0
dht11==0.1.0